ch5/src/test/java/com/springinaction/ch5/web/SpittleControllerTest.java
package com.springinaction.ch5.web;

import com.springinaction.ch5.business.data.SpittleRepository;
import com.springinaction.ch5.business.model.Spittle;
import org.junit.Test;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;
import org.springframework.test.web.servlet.ResultMatcher;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
import org.springframework.test.web.servlet.result.ModelResultMatchers;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import org.springframework.web.servlet.view.InternalResourceView;

import java.util.ArrayList;
import java.util.Date;
import java.util.List;

import static org.hamcrest.core.IsCollectionContaining.hasItems;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

public class SpittleControllerTest {

    @Test
    public void shouldShowRecentSpittles() throws Exception {
        List<Spittle> spittles = createSpittles(20);

        SpittleRepository spittleRepository = mock(SpittleRepository.class);
        when(spittleRepository.findSpittleList(Long.MAX_VALUE, 20)).thenReturn(spittles);

        SpittleController controller = new SpittleController(spittleRepository);

        final MockMvc mockMvc = MockMvcBuilders.standaloneSetup(controller)
                .setSingleView(new InternalResourceView("/WEB-INF/views/spittles.jsp"))
                .build();

        mockMvc.perform(MockMvcRequestBuilders.get("/spittles"))
                .andExpect(MockMvcResultMatchers.view().name("spittles"))
                .andExpect(MockMvcResultMatchers.model().attributeExists("spittleList"))
                .andExpect(MockMvcResultMatchers.model().attribute("spittleList", hasItems(spittles.toArray())));
    }

    private List<Spittle> createSpittles(int count) {
        List<Spittle> spittles = new ArrayList<Spittle>(count);
        for (int i = 0 ; i < count; i++) {
            spittles.add(new Spittle("spittle " + i, new Date()));
        }

        return spittles;
    }
}
